add float, and director groups

+++ admin/sql/install.mysql.utf8.sql
@@ -120,8 +120,8 @@ CREATE TABLE IF NOT EXISTS `#__workhour` (
 	`approve_comment` VARCHAR(250),
 	`start_datetime` DATETIME NOT NULL DEFAULT '0000-00-00 00:00:00',
 	`complete_datetime` DATETIME NOT NULL DEFAULT '0000-00-00 00:00:00',
-	`hours_submitted` INT(4) NOT NULL DEFAULT '0',
-	`hours_approved` INT(4) NOT NULL DEFAULT '0',
+	`hours_submitted` FLOAT(4) NOT NULL DEFAULT '0',
+	`hours_approved` FLOAT(4) NOT NULL DEFAULT '0',
 	`approved` TINYINT(1) NOT NULL DEFAULT '0',
 	`approved_by` INT(10) NOT NULL DEFAULT '0',
     PRIMARY KEY (`id`)
